+++ bin/flutter_figma_theme_generator.dart
@@ -13,8 +13,11 @@ import 'package:path/path.dart';
 
 Future<void> main(List<String> args) async {
   final argParser = ArgParser()
-    ..addOption('path', help: 'Override the default theme folder path. This value will be used instead of the default OR what you have configured in pubspec.yaml')
-    ..addFlag('help', help: 'Displays this help screen', defaultsTo: false, negatable: false);
+    ..addOption('path',
+        help:
+            'Override the default theme folder path. This value will be used instead of the default OR what you have configured in pubspec.yaml')
+    ..addFlag('help',
+        help: 'Displays this help screen', defaultsTo: false, negatable: false);
 
   final results = argParser.parse(args);
   if (results['help']) {
@@ -24,7 +27,8 @@ Future<void> main(List<String> args) async {
 
   final pubspecYaml = File(join(Directory.current.path, 'pubspec.yaml'));
   if (!pubspecYaml.existsSync()) {
-    throw Exception('This program should be run from the root of a flutter/dart project');
+    throw Exception(
+        'This program should be run from the root of a flutter/dart project');
   }
   final pubspecContent = pubspecYaml.readAsStringSync();
   final pubspecConfig = PubspecConfig(pubspecContent);
@@ -38,26 +42,41 @@ Future<void> main(List<String> args) async {
   }
 
   if (!configFolder.existsSync()) {
-    throw Exception('This program requires a config folder. Configuration is set to `$configPath`');
+    throw Exception(
+        'This program requires a config folder. Configuration is set to `$configPath`');
   }
-  final themeFiles = configFolder.listSync(recursive: true).whereType<File>().where((e) => e.path.endsWith('.json'));
+  final themeFiles = configFolder
+      .listSync(recursive: true)
+      .whereType<File>()
+      .where((e) => e.path.endsWith('.json'));
   final fontGenerator = FontGenerator();
   final nonThemeGenerators = [fontGenerator, ColorGenerator()];
   final contents = await _readFiles(themeFiles);
   var defaultTheme = pubspecConfig.defaultTheme ??
-      contents.entries.firstWhere((element) => nonThemeGenerators.any((e) => e.matchesSchema(element.value)), orElse: () => throw Exception('Could not find any theme')).key;
+      contents.entries
+          .firstWhere(
+              (element) =>
+                  nonThemeGenerators.any((e) => e.matchesSchema(element.value)),
+              orElse: () => throw Exception('Could not find any theme'))
+          .key;
   final hasFontTheme = contents.values.any(fontGenerator.matchesSchema);
   final generatedContents = <GeneratedContent>[];
   final warnings = <String>[];
   if (!contents.containsKey(defaultTheme)) {
-    warnings.add('Could not find the default theme `$defaultTheme` in the theme folder. Defaulting to the first theme found. Found: ${contents.keys.join(', ')}');
+    warnings.add(
+        'Could not find the default theme `$defaultTheme` in the theme folder. Defaulting to the first theme found. Found: ${contents.keys.join(', ')}');
     defaultTheme = contents.entries.first.key;
   }
 
-  generatedContents.add(await _processThemeFile(contents[defaultTheme]!, defaultTheme, pubspecConfig));
-  generatedContents.addAll(await Future.wait(contents.entries.where((element) => element.key != defaultTheme).map((e) => _processThemeFile(e.value, e.key, pubspecConfig))));
-  final generatedInstances = generatedContents.map((e) => e.themeInstanceName).whereType<String>();
-  generatedContents.add(await _generateThemeFile(generatedInstances, pubspecConfig, hasFontTheme));
+  generatedContents.add(await _processThemeFile(
+      contents[defaultTheme]!, defaultTheme, pubspecConfig));
+  generatedContents.addAll(await Future.wait(contents.entries
+      .where((element) => element.key != defaultTheme)
+      .map((e) => _processThemeFile(e.value, e.key, pubspecConfig))));
+  final generatedInstances =
+      generatedContents.map((e) => e.themeInstanceName).whereType<String>();
+  generatedContents.add(await _generateThemeFile(
+      generatedInstances, pubspecConfig, hasFontTheme));
 
   final allWarnings = generatedContents.expand((e) => e.warnings).toSet();
   allWarnings.addAll(warnings);
@@ -71,8 +90,10 @@ Future<void> main(List<String> args) async {
   }
 }
 
-Future<GeneratedContent> _generateThemeFile(Iterable<String> generatedInstances, PubspecConfig pubspecConfig, bool hasFontTheme) async {
-  final generatedTheme = CurrentThemeGenerator.generateTheme(generatedInstances, pubspecConfig, hasFontTheme);
+Future<GeneratedContent> _generateThemeFile(Iterable<String> generatedInstances,
+    PubspecConfig pubspecConfig, bool hasFontTheme) async {
+  final generatedTheme = CurrentThemeGenerator.generateTheme(
+      generatedInstances, pubspecConfig, hasFontTheme);
   await _createFile(generatedTheme);
   return generatedTheme;
 }
@@ -93,14 +114,19 @@ Future<void> _createFile(GeneratedContent generatedTheme) async {
   }));
 }
 
-Future<Map<String, Map<String, dynamic>>> _readFiles(Iterable<File> themeFiles) async {
+Future<Map<String, Map<String, dynamic>>> _readFiles(
+    Iterable<File> themeFiles) async {
   final contents = <String, Map<String, dynamic>>{};
-  await Future.wait(themeFiles.map((file) async => contents[file.path.substring(file.path.lastIndexOf('/') + 1)] = jsonDecode(await file.readAsString())));
+  await Future.wait(themeFiles.map((file) async =>
+      contents[file.path.substring(file.path.lastIndexOf('/') + 1)] =
+          jsonDecode(await file.readAsString())));
   return contents;
 }
 
-Future<GeneratedContent> _processThemeFile(Map<String, dynamic> content, String path, PubspecConfig pubspecConfig) async {
-  final generatedTheme = ThemeGenerator.generateTheme(content, path, pubspecConfig);
+Future<GeneratedContent> _processThemeFile(Map<String, dynamic> content,
+    String path, PubspecConfig pubspecConfig) async {
+  final generatedTheme =
+      ThemeGenerator.generateTheme(content, path, pubspecConfig);
   await _createFile(generatedTheme);
   return generatedTheme;
 }