+++ lib/config/pubspec_config.dart
@@ -18,9 +18,13 @@ class PubspecConfig {
     }
     final projectName = doc['name'];
     if (projectName == null || projectName.isEmpty) {
-      throw Exception('Could not parse the pubspec.yaml, project name not found');
+      throw Exception(
+          'Could not parse the pubspec.yaml, project name not found');
     }
-    fonts.addAll(doc['flutter']?['fonts']?.map<String>((e) => e['family'].toString()).toList() ?? []);
+    fonts.addAll(doc['flutter']?['fonts']
+            ?.map<String>((e) => e['family'].toString())
+            .toList() ??
+        []);
 
     this.projectName = projectName;
     final config = doc['flutter_figma_theme_generator'];