+++ lib/generators/font_generator.dart
@@ -6,10 +6,12 @@ import 'package:flutter_figma_theme_generator/utils/case_utils.dart';
 
 class FontGenerator extends BaseGenerator {
   @override
-  bool matchesSchema(Map<String, dynamic> schema) => schema['Typography'] != null;
+  bool matchesSchema(Map<String, dynamic> schema) =>
+      schema['Typography'] != null;
 
   @override
-  GeneratedContent generate(Map<String, dynamic> schema, PubspecConfig pubspecConfig) {
+  GeneratedContent generate(
+      Map<String, dynamic> schema, PubspecConfig pubspecConfig) {
     final typography = schema['Typography'] as Map<String, dynamic>;
     final textStyles = <String, TextStyle>{};
     final files = <String, String>{};
@@ -19,29 +21,52 @@ class FontGenerator extends BaseGenerator {
       final json = entry.value as Map<String, dynamic>;
       textStyles.addAll(_generateTextStyles(entry.key, json, typography));
     }
-    final fontFamilies = textStyles.entries.map((e) => e.value.fontFamily).toSet()..removeWhere(pubspecConfig.fonts.contains);
-    final warnings = fontFamilies.map((fontFamily) => 'Warning: Font family "$fontFamily" is not defined in pubspec.yaml').toList();
+    final fontFamilies = textStyles.entries
+        .map((e) => e.value.fontFamily)
+        .toSet()
+      ..removeWhere(pubspecConfig.fonts.contains);
+    final warnings = fontFamilies
+        .map((fontFamily) =>
+            'Warning: Font family "$fontFamily" is not defined in pubspec.yaml')
+        .toList();
 
     var textStyleFile = '''import 'package:flutter/material.dart';\n\n''';
-    textStyleFile += 'class ${pubspecConfig.projectName.upperCamelCase}TextTheme {\n';
+    textStyleFile +=
+        'class ${pubspecConfig.projectName.upperCamelCase}TextTheme {\n';
     for (final textStyle in textStyles.entries) {
-      textStyleFile += '  final TextStyle ${textStyle.key};\n';
+      textStyleFile += '  late final TextStyle ${textStyle.key};\n';
     }
-    textStyleFile += '\n  const ${pubspecConfig.projectName.upperCamelCase}TextTheme()\n      : ';
-    textStyleFile += textStyles.entries.map((textStyle) => '${textStyle.key} = const ${textStyle.value.toString()}').join(',\n        ');
-    textStyleFile += ';\n}\n';
+    textStyleFile +=
+        '\n  ${pubspecConfig.projectName.upperCamelCase}TextTheme({\n';
+    textStyleFile += '    TextStyle? allOverrides,\n';
+    textStyleFile += textStyles.entries
+        .map((textStyle) => '    TextStyle? ${textStyle.key}Overrides,\n')
+        .join();
+    textStyleFile += '  }) {\n';
+    textStyleFile += textStyles.entries
+        .map((textStyle) =>
+            '    ${textStyle.key} = ${textStyle.value.stringWithOverrides('${textStyle.key}Overrides')};\n')
+        .join('');
+    textStyleFile += '  }\n';
+    textStyleFile += '}\n';
 
     files['${pubspecConfig.projectName.snakeCase}_text_theme'] = textStyleFile;
 
     return GeneratedContent(files, warnings);
   }
 
-  TextStyle _generateTextStyle(Map<String, dynamic> data, Map<String, dynamic> typography) {
-    final fontFamily = _fromDataOrTypography(data['fontFamily'] as String?, typography);
-    final fontWeight = _fromDataOrTypography(data['fontWeight'] as String?, typography);
-    final fontSizeData = _fromDataOrTypography(data['fontSize'] as String?, typography);
-    final lineHeight = _fromDataOrTypography(data['lineHeight'] as String?, typography);
-    final letterSpacing = _fromDataOrTypography(data['letterSpacing'] as String?, typography);
+  TextStyle _generateTextStyle(
+      Map<String, dynamic> data, Map<String, dynamic> typography) {
+    final fontFamily =
+        _fromDataOrTypography(data['fontFamily'] as String?, typography);
+    final fontWeight =
+        _fromDataOrTypography(data['fontWeight'] as String?, typography);
+    final fontSizeData =
+        _fromDataOrTypography(data['fontSize'] as String?, typography);
+    final lineHeight =
+        _fromDataOrTypography(data['lineHeight'] as String?, typography);
+    final letterSpacing =
+        _fromDataOrTypography(data['letterSpacing'] as String?, typography);
 
     final fontSize = double.tryParse(fontSizeData ?? '');
 
@@ -57,20 +82,32 @@ class FontGenerator extends BaseGenerator {
   String? _fromDataOrTypography(String? data, Map<String, dynamic> typography) {
     if (data == null) return null;
     if (data.contains(RegExp(r'(\/|\*|round\(.+\)|\(.+\))'))) {
-      return RegExp(r'({[^{}]+}|\d+|round\(.+\)|\(.+\))').allMatches(data).fold(1.0, (previousValue, element) {
+      return RegExp(r'({[^{}]+}|\d+|round\(.+\)|\(.+\))')
+          .allMatches(data)
+          .fold(1.0, (previousValue, element) {
         String operator;
         if (element.start == 0) {
           if (data.startsWith('round(') && data.endsWith(')')) {
-            return double.tryParse(_fromDataOrTypography(data.substring(6, data.length - 1), typography) ?? '')?.roundToDouble() ?? 1.0;
+            return double.tryParse(_fromDataOrTypography(
+                            data.substring(6, data.length - 1), typography) ??
+                        '')
+                    ?.roundToDouble() ??
+                1.0;
           } else if (data.startsWith('(') && data.endsWith(')')) {
-            return double.tryParse(_fromDataOrTypography(data.substring(1, data.length - 1), typography) ?? '') ?? 1.0;
+            return double.tryParse(_fromDataOrTypography(
+                        data.substring(1, data.length - 1), typography) ??
+                    '') ??
+                1.0;
           } else {
             operator = '*';
           }
         } else {
           operator = data.substring(element.start - 1, element.start);
         }
-        final nextValue = double.tryParse(_fromDataOrTypography(data.substring(element.start, element.end), typography) ?? '') ?? 1;
+        final nextValue = double.tryParse(_fromDataOrTypography(
+                    data.substring(element.start, element.end), typography) ??
+                '') ??
+            1;
         if (operator == '/') {
           return previousValue / nextValue;
         } else {
@@ -91,14 +128,17 @@ class FontGenerator extends BaseGenerator {
   double? _getLetterSpacing(String? letterSpacing) {
     if (letterSpacing == null) return null;
     if (letterSpacing.endsWith('%')) {
-      final percentage = double.tryParse(letterSpacing.substring(0, letterSpacing.length - 1));
+      final percentage =
+          double.tryParse(letterSpacing.substring(0, letterSpacing.length - 1));
       return percentage != null ? percentage / 100 : null;
     }
     return double.tryParse(letterSpacing);
   }
 
   double? _getHeight(String? lineHeight, double? fontSize) {
-    if (lineHeight == null || lineHeight == 'AUTO' || fontSize == null) return null;
+    if (lineHeight == null || lineHeight == 'AUTO' || fontSize == null) {
+      return null;
+    }
     final height = double.tryParse(lineHeight);
     return height == null ? null : height / fontSize;
   }
@@ -118,13 +158,17 @@ class FontGenerator extends BaseGenerator {
     }
   }
 
-  Map<String, TextStyle> _generateTextStyles(String key, Map<String, dynamic> data, Map<String, dynamic> typography) {
+  Map<String, TextStyle> _generateTextStyles(
+      String key, Map<String, dynamic> data, Map<String, dynamic> typography) {
     final textStyles = <String, TextStyle>{};
     for (final entry in data.entries) {
       if (entry.key == 'value' && entry.value is Map<String, dynamic>) {
-        textStyles[key.camelCase] = _generateTextStyle(entry.value as Map<String, dynamic>, typography);
-      } else if (!_isNotFont(entry.key) && entry.value is Map<String, dynamic>) {
-        textStyles.addAll(_generateTextStyles('${key}_${entry.key}'.camelCase, entry.value as Map<String, dynamic>, typography));
+        textStyles[key.camelCase] =
+            _generateTextStyle(entry.value as Map<String, dynamic>, typography);
+      } else if (!_isNotFont(entry.key) &&
+          entry.value is Map<String, dynamic>) {
+        textStyles.addAll(_generateTextStyles('${key}_${entry.key}'.camelCase,
+            entry.value as Map<String, dynamic>, typography));
       }
     }
     return textStyles;