+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Added `textThemeBuilder` method to allow for setting default colors and/or overrides for all or specific text styles
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Create themes from figma tokens